Fully automated staging and deployment
My tools now allow me to do fully automated testing and deployment.
Unfortunately, several of these tools are not yet in-tree -- the virtual
machine that runs Windows; the applescript that automatically fires up
IE in the VM and runs the tests, etc…  Perhaps later it will be possible
to include these scripts?  I have also split the unit tests and the
tests for the link structure of index.php. When IE was visiting a .js
file, it was popping up a prompt to download the file, which is
different from Firefox and Chrome. Therefore, I only test links in
Firefox but run the more expansive unit tests in all browsers.  I have
also fixed some bugs.