Ensure proper binding for dynamically generated functions.
Many places in numeric.js create new functions using the Function(...)
constructor.  Such functions do not have access to the local scope.
Therefore they bind to the global "numeric" object, which might not exist,
or might belong to a different version of numeric.js, or might be something
different altogether.  Since we are careful enough to have a local variable
in most places, it appears prudent to make sure that we pass this to
functions created from source strings as well.